fix : 상단 필터부분 분리
